1.Check if an employee already exists as a given job before assigning it a new one 2. Fix office update

+++ backend/src/main/java/com/nbu/logisticcompany/services/CourierServiceImpl.java
@@ -3,8 +3,11 @@ package com.nbu.logisticcompany.services;
 import com.nbu.logisticcompany.entities.Courier;
 import com.nbu.logisticcompany.entities.OfficeEmployee;
 import com.nbu.logisticcompany.entities.User;
+import com.nbu.logisticcompany.exceptions.DuplicateEntityException;
+import com.nbu.logisticcompany.exceptions.EntityNotFoundException;
 import com.nbu.logisticcompany.repositories.interfaces.CourierRepository;
 import com.nbu.logisticcompany.services.interfaces.CourierService;
+import com.nbu.logisticcompany.services.interfaces.OfficeEmployeeService;
 import com.nbu.logisticcompany.utils.Action;
 import com.nbu.logisticcompany.utils.ValidationUtil;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,10 +22,12 @@ import static com.nbu.logisticcompany.utils.ValidationUtil.validateAdminAction;
 public class CourierServiceImpl implements CourierService {
 
     private final CourierRepository courierRepository;
+    private final OfficeEmployeeService officeEmployeeService;
 
     @Autowired
-    public CourierServiceImpl(CourierRepository courierRepository) {
+    public CourierServiceImpl(CourierRepository courierRepository, OfficeEmployeeService _officeEmployeeService) {
         this.courierRepository = courierRepository;
+        officeEmployeeService = _officeEmployeeService;
     }
 
     @Override
@@ -71,8 +76,20 @@ public class CourierServiceImpl implements CourierService {
     @Override
     public void makeOfficeEmployee(int courierId, int officeId, User updater) {
         validateAdminAction(updater, OfficeEmployee.class, Action.UPDATE);
+        if (doesUserExistAsOfficeEmployee(courierId)){
+            throw new DuplicateEntityException(OfficeEmployee.class.getSimpleName(), "id", String.valueOf(courierId));
+        }
         courierRepository.removeUserFromCouriers(courierId);
         courierRepository.makeOfficeEmployee(courierId, officeId);
+    }
+
+    private boolean doesUserExistAsOfficeEmployee(int courierId) {
+        try {
+            officeEmployeeService.getById(courierId);
+            return true;
+        } catch (EntityNotFoundException ex) {
+            return false;
+        }
     }
 
 }
